+++ src/main/kotlin/game/game/skill/farming/Patch.kt
@@ -1,8 +1,8 @@
 package game.skill.farming
 
 /**
- *
  * @author lare96
  */
+@Deprecated("Use FarmingPatch instead")
 class Patch(val type: PatchType, val display: Int, val state: Int, private var stage: Int) {
 }
