+++ src/main/kotlin/game/game/skill/farming/PatchType.kt
@@ -4,6 +4,7 @@ package game.skill.farming
  *
  * @author lare96
  */
+@Deprecated("Use instead subclasses of FarmingPatch")
 enum class PatchType(val cycleMinutes: Int = 1) {
     ALLOTMENT(cycleMinutes = 10),
     FLOWER,
